Fix building process for picard 2.10
Since 2.8.5 (last published image), python3.7 became
required which in turn required using a more modern
ubuntu base.  22.04 was choosen in light of "newer is
better for audio tools" and since it removes the need
to use a secondary ppa.

The rest of the changes are either correcting package names
that have changed, or correcting for a file having moved in a repository.

Signed-off-by: Brian Harring <ferringb@gmail.com>